Added user navigation hook
