+++ tweeter-web-starter/tweeter-web/src/components/statusItem/StatusItem.tsx
@@ -1,51 +1,15 @@
 import { Link } from "react-router-dom";
 import Post from "./Post";
-import { AuthToken, User, FakeData, Status } from "tweeter-shared";
-import { useContext } from "react";
-import { UserInfoContext } from "../userInfo/UserInfoProvider";
-import useToastListener from "../toaster/ToastListenerHook";
+import { Status } from "tweeter-shared";
+import useUserNavigation from "../userInfo/UserNavigationHook";
 
 interface Props {
   value: Status;
 }
 
 const StatusItem = (props: Props) => {
-  const { displayErrorMessage } = useToastListener();
-  const { setDisplayedUser, currentUser, authToken } =
-    useContext(UserInfoContext);
-  const navigateToUser = async (event: React.MouseEvent): Promise<void> => {
-    event.preventDefault();
-
-    try {
-      const alias = extractAlias(event.target.toString());
-
-      const user = await getUser(authToken!, alias);
-
-      if (!!user) {
-        if (currentUser!.equals(user)) {
-          setDisplayedUser(currentUser!);
-        } else {
-          setDisplayedUser(user);
-        }
-      }
-    } catch (error) {
-      displayErrorMessage(`Failed to get user because of exception: ${error}`);
-    }
-  };
-
-  const extractAlias = (value: string): string => {
-    const index = value.indexOf("@");
-    return value.substring(index);
-  };
-
-  const getUser = async (
-    authToken: AuthToken,
-    alias: string
-  ): Promise<User | null> => {
-    // TODO: Replace with the result of calling server
-    return FakeData.instance.findUserByAlias(alias);
-  };
-
+  const { navigateToUser } = useUserNavigation();
+  
   return (
     <div className="col bg-light mx-0 px-0">
       <div className="container px-0">
